Resume_Manager/requirements.txt
flask
python-docx
pymupdf
transformers
accelerate
torch
huggingface-hub
google-generativeai